+++ src/main/java/cart/controller/MemberArgumentResolver.java
@@ -1,8 +1,8 @@
 package cart.controller;
 
 import cart.dao.MemberDao;
-import cart.domain.Member;
-import cart.domain.Password;
+import cart.domain.member.Member;
+import cart.domain.member.Password;
 import cart.exception.AuthenticationException;
 import org.apache.tomcat.util.codec.binary.Base64;
 import org.springframework.core.MethodParameter;
