+++ src/main/java/cart/dao/MemberDao.java
@@ -1,6 +1,6 @@
 package cart.dao;
 
-import cart.domain.Member;
+import cart.domain.member.Member;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
