+++ src/test/java/cart/domain/EmailTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.member.Email;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
