+++ src/test/java/cart/domain/ImageUrlTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.product.ImageUrl;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
